Moving code in right files
