+++ src/index.js
@@ -1,4 +1,5 @@
 import './sass/main.scss';
+import './js/dom/pagination'
 import './js/dom/show-all-films';
 import './js/dom/show-searched-films';
 import './js/dom/home-library-page-switch'
